backend/requirements.txt
asgiref==3.3.4
Django>=3.2,<3.3.0
djangorestframework>=3.12.4,<3.13.0
django-cors-headers==3.7.0
flake8>=3.6.0,<3.7.0
pytz==2021.1
Pillow==8.2.0
djangorestframework-simplejwt==4.6.0
gunicorn==20.1.0
whitenoise==5.2.0